python-fx/requirements.txt
flask
requests
psycopg2-binary
gunicorn